Move resize related methods from FS to DeviceFormat
This includes update_min_size, do_resize, current_size,
target_size, _get_target_size, _set_target_size, _get_size methods
and properties.
